+++ assembly/src/assembler/mod.rs
@@ -7,7 +7,7 @@ use vm_core::{
     crypto::hash::RpoDigest,
     debuginfo::SourceSpan,
     mast::{DecoratorId, MastNodeId},
-    DecoratorList, Felt, Kernel, Operation, Program,
+    DecoratorList, Felt, Kernel, Operation, Program, WORD_SIZE,
 };
 
 use crate::{
@@ -542,19 +542,21 @@ impl Assembler {
     ) -> Result<Procedure, Report> {
         // Make sure the current procedure context is available during codegen
         let gid = proc_ctx.id();
+
         let num_locals = proc_ctx.num_locals();
 
         let wrapper_proc = self.module_graph.get_procedure_unsafe(gid);
         let proc = wrapper_proc.unwrap_ast().unwrap_procedure();
         let proc_body_id = if num_locals > 0 {
-            // for procedures with locals, we need to update fmp register before and after the
-            // procedure body is executed. specifically:
+            // For procedures with locals, we need to update fmp register before and after the
+            // procedure body is executed. Specifically:
             // - to allocate procedure locals we need to increment fmp by the number of locals
-            // - to deallocate procedure locals we need to decrement it by the same amount
-            let num_locals = Felt::from(num_locals);
+            //   (rounded up to the word size), and
+            // - to deallocate procedure locals we need to decrement it by the same amount.
+            let locals_frame = Felt::from(num_locals.next_multiple_of(WORD_SIZE as u16));
             let wrapper = BodyWrapper {
-                prologue: vec![Operation::Push(num_locals), Operation::FmpUpdate],
-                epilogue: vec![Operation::Push(-num_locals), Operation::FmpUpdate],
+                prologue: vec![Operation::Push(locals_frame), Operation::FmpUpdate],
+                epilogue: vec![Operation::Push(-locals_frame), Operation::FmpUpdate],
             };
             self.compile_body(proc.iter(), &mut proc_ctx, Some(wrapper), mast_forest_builder)?
         } else {