+++ test-utils/src/test_builders.rs
@@ -18,7 +18,7 @@
 macro_rules! build_op_test {
     ($op_str:expr) => {{
         let source = format!("
-proc.truncate_stack.1
+proc.truncate_stack.4
     loc_storew.0 dropw movupw.3
     sdepth neq.16
     while.true
@@ -35,7 +35,7 @@ begin {} exec.truncate_stack end",
     }};
     ($op_str:expr, $($tail:tt)+) => {{
         let source = format!("
-proc.truncate_stack.1
+proc.truncate_stack.4
     loc_storew.0 dropw movupw.3
     sdepth neq.16
     while.true